Implement styled lead report app
