+++ frontend/RealtorInterface/LeadReports/Dockerfile
@@ -5,6 +5,6 @@ RUN npm install && npm run build
 
 FROM nginx:alpine
 COPY frontend/RealtorInterface/LeadReports/nginx.conf /etc/nginx/conf.d/default.conf
-COPY --from=build /app/reports/dist /usr/share/nginx/html/console/reports
+COPY --from=build /app/reports/dist /usr/share/nginx/html/console/leadreport
 EXPOSE 80
 CMD ["nginx", "-g", "daemon off;"]
