+++ frontend/RealtorInterface/LeadReports/nginx.conf
@@ -3,8 +3,8 @@ server {
     location /api/ {
         proxy_pass http://api:3000/api/;
     }
-    location /console/reports/ {
-        alias /usr/share/nginx/html/console/reports/;
-        try_files $uri $uri/ /console/reports/index.html;
+    location /console/leadreport/ {
+        alias /usr/share/nginx/html/console/leadreport/;
+        try_files $uri $uri/ /console/leadreport/index.html;
     }
 }
