+++ frontend/RealtorInterface/LeadReports/package.json
@@ -10,11 +10,15 @@
   "dependencies": {
     "@types/react": "^19.1.6",
     "@types/react-dom": "^19.1.5",
+    "lucide-react": "^0.514.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^4.0.0",
+    "autoprefixer": "^10.4.16",
+    "postcss": "^8.4.34",
+    "tailwindcss": "^3.4.0",
     "vite": "^5.0.0"
   }
 }
