+++ frontend/RealtorInterface/LeadReports/src/main.css
@@ -1,0 +1,3 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
