+++ frontend/RealtorInterface/LeadReports/src/main.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.jsx';
+import './main.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
